project_evolution_part2/requirements.txt
pandas
psycopg2-binary
matplotlib
seaborn
python-dotenv
requests
reportlab
great_expectations>=0.17.20
pydantic<2